+++ methods.go
@@ -21,6 +21,7 @@ type Method string
 type Event string
 
 const (
+	MethodConnectorConnect Method = "sql.connector.connect"
 	MethodConnPing         Method = "sql.conn.ping"
 	MethodConnExec         Method = "sql.conn.exec"
 	MethodConnQuery        Method = "sql.conn.query"